# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Catálogo de Ferias - API canónica
  version: 1.0.0
  description: Contrato HTTP estable del backend de Unidades Productivas y ferias.
servers:
  - url: http://localhost:5000/api
security:
  - bearerAuth: []
...
paths:
...
  /admin/users/{userId}:
    parameters: [{$ref: '#/components/parameters/UserId'}]
    get:
      tags: [Usuarios]
      summary: Consultar usuario
      responses: {'200': {description: Usuario}, '404': {$ref: '#/components/responses/NotFound'}}
    patch:
      tags: [Usuarios]
      summary: Editar identidad o rol
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses: {'200': {description: Usuario actualizado}, '409': {$ref: '#/components/responses/Conflict'}}
    delete:
      tags: [Usuarios]
      summary: Eliminar usuario lógicamente
      responses: {'204': {description: Eliminado}}
  /admin/users/{userId}/status:
    patch:
      tags: [Usuarios]
      summary: Cambiar estado
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {$ref: '#/components/requestBodies/JsonBody'}
      responses: {'200': {description: Estado actualizado}}
  /admin/users/{userId}/unlock:
    post:
      tags: [Usuarios]
      summary: Desbloquear usuario
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses: {'200': {description: Usuario desbloqueado}}
  /admin/users/{userId}/restore:
    post:
      tags: [Usuarios]
      summary: Restaurar usuario
      parameters: [{$ref: '#/components/parameters/UserId'}]
      responses: {'200': {description: Usuario restaurado}}
...
  /admin/audits/{auditId}:
    get:
      tags: [Auditoría]
      summary: Consultar auditoría
      parameters: [{$ref: '#/components/parameters/AuditId'}]
      responses: {'200': {description: Evento auditado}, '404': {$ref: '#/components/responses/NotFound'}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId: {name: userId, in: path, required: true, schema: {type: string, format: uuid}}
...
    AuditId: {name: auditId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
...
  responses:
...
    NotFound: {description: Recurso no encontrado}
    Conflict: {description: Conflicto con el estado o una regla de negocio}
...